answers tableのマイグレーション
